[feat] 로그인 구현

+++ src/main/kotlin/com/example/onlineshoping/project/domain/dto/request/CreateBuyerRequest.kt
@@ -5,5 +5,5 @@ data class CreateBuyerRequest(
     val password:String,
     val name:String,
     val phoneNumber:String,
-    val balance:Long
+    val balance:Int
 )
